requirements.txt
flask==2.3.3
flask-login==0.6.2
pymongo==4.5.0
python-dotenv==1.0.0
dlib
face-recognition
face-recognition-models
numpy==1.24.3
opencv-python-headless==4.7.0.72
Pillow==10.0.0
openpyxl==3.1.2